Add error reporting

+++ DiskSmartInfo.functions.ps1
@@ -11,43 +11,92 @@ function Get-DiskSmartInfo
         [switch]$SilenceIfNotInWarningOrCriticalState
     )
 
+    $Script:ErrorCreatingCimSession = @()
+    $Script:ErrorAccessingCimSession = @()
+    $Script:ErrorAccessingClass = @()
+
+    # ComputerName
     if ($ComputerName)
     {
         try
         {
-            $cimSessions = New-CimSession -ComputerName $ComputerName
-
+            # $cimSessions = New-CimSession -ComputerName $ComputerName -ErrorVariable ErrorCreatingCimSession -ErrorAction SilentlyContinue
+            $cimSessions = New-CimSession -ComputerName $ComputerName -ErrorVariable Script:ErrorCreatingCimSession -ErrorAction SilentlyContinue
             foreach ($cim in $cimSessions)
             {
                 inGetDiskSmartInfo `
-                    -Session $cim `
-                    -ShowConvertedData:$ShowConvertedData `
-                    -CriticalAttributesOnly:$CriticalAttributesOnly `
-                    -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
-            }
-        }
-        finally
-        {
-            Remove-CimSession -CimSession $cimSessions
-        }
-    }
-    elseif ($CimSession)
-    {
-        foreach ($cim in $CimSession)
-        {
-            inGetDiskSmartInfo `
                 -Session $cim `
                 -ShowConvertedData:$ShowConvertedData `
                 -CriticalAttributesOnly:$CriticalAttributesOnly `
                 -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
+            }
         }
+        finally
+        {
+            if ($cimSessions)
+            {
+                Remove-CimSession -CimSession $cimSessions
+            }
+        }
+
+        # foreach ($eccs in $ErrorCreatingCimSession)
+        # foreach ($eccs in $Errors)
+        # {
+        #     Write-Error -Message "Can't connect to '$($eccs.OriginInfo.PSComputerName)'."
+        # }
     }
+
+    # CimSession
+    elseif ($CimSession)
+    {
+        foreach ($cim in $CimSession)
+        {
+            if ($cim.TestConnection())
+            {
+                inGetDiskSmartInfo `
+                -Session $cim `
+                -ShowConvertedData:$ShowConvertedData `
+                -CriticalAttributesOnly:$CriticalAttributesOnly `
+                -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
+            }
+            else
+            {
+                $Script:ErrorAccessingCimSession += $cim.ComputerName
+            }
+        }
+        # foreach ($eacs in $ErrorAccessingCimSession)
+        # {
+        #     Write-Error -Message "Can't connect to '$eacs'."
+        # }
+    }
+
+    # Localhost
     else
     {
         inGetDiskSmartInfo `
             -ShowConvertedData:$ShowConvertedData `
             -CriticalAttributesOnly:$CriticalAttributesOnly `
             -SilenceIfNotInWarningOrCriticalState:$SilenceIfNotInWarningOrCriticalState
+    }
+
+    # foreach ($eac in $Script:ErrorAccessingClass)
+    # {
+    #     Write-Error -Message $eac
+    # }
+
+    foreach ($e in $Script:ErrorCreatingCimSession)
+    {
+        Write-Error -Message "ComputerName: ""$($e.OriginInfo.PSComputerName)"". $($e.Exception.Message)"
+    }
+    foreach ($e in $Script:ErrorAccessingCimSession)
+    {
+        Write-Error -Message "ComputerName: ""$e"". The WinRM client cannot process the request because the CimSession cannot be accessed."
+        # Write-Error -Message "ComputerName: ""$($e.OriginInfo.PSComputerName)"". $($e.Exception.Message)"
+    }
+    foreach ($e in $Script:ErrorAccessingClass)
+    {
+        Write-Error -Message "ComputerName: ""$($e.ComputerName)"", Protocol: $($e.Protocol). $($e.ErrorObject.Exception.Message)"
+        # $Script:ErrorAccessingClass += "ComputerName: ""$($Session.ComputerName)"", Protocol: $($Session.Protocol)`n$($_.Exception.Message)"
     }
 }
 
@@ -75,9 +124,18 @@ function inGetDiskSmartInfo
         $parameters.Add('CimSession', $Session)
     }
 
-    $disksInfo = Get-CimInstance -Namespace $namespaceWMI -ClassName $classSMARTData @parameters
-    $disksThresholds = Get-CimInstance -Namespace $namespaceWMI -ClassName $classThresholds @parameters
-    $diskDrives = Get-CimInstance -ClassName $classDiskDrive @parameters
+    try
+    {
+        $disksInfo = Get-CimInstance -Namespace $namespaceWMI -ClassName $classSMARTData @parameters  -ErrorAction Stop
+        $disksThresholds = Get-CimInstance -Namespace $namespaceWMI -ClassName $classThresholds @parameters
+        $diskDrives = Get-CimInstance -ClassName $classDiskDrive @parameters
+    }
+    catch
+    {
+        # $Script:ErrorAccessingClass += "ComputerName: ""$($Session.ComputerName)"", Protocol: $($Session.Protocol)`n$($_.Exception.Message)"
+        $Script:ErrorAccessingClass += @{ComputerName = $Session.ComputerName; Protocol = $Session.Protocol; ErrorObject = $_}
+        continue
+    }
 
     foreach ($diskInfo in $disksInfo)
     {
